requirements.txt
# external requirements
python-dotenv==1.0.1
chromadb==0.4.24
langchain==0.1.11
langchain-community==0.0.27
pymupdf==1.23.26
sentence-transformers==2.5.1
tensorflow==2.16.1
gradio==4.21.0
pandas==2.2.1
streamlit==1.32.2
